+++ backend/src/auth/types/index.ts
@@ -1,1 +1,2 @@
 export * from './github-profile.interface';
+export * from './payload.interface';
